% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{net_noise}
\alias{net_noise}
\title{This function takes a network and either adds or deletes a percentage of
edges randomly}
\usage{
net_noise(
  network,
  mode = "add",
  perc = 0.1,
  source = "source",
  target = "target",
  seed = 42
)
}
\arguments{
\item{network}{Network to modify.}

\item{mode}{Whether to add ('add') or delete ('del') edges.}

\item{perc}{Percentage of edges to modify per regulon.}

\item{source}{Network's source column.}

\item{target}{Network's target genes column.}

\item{seed}{An integer to set the RNG state for random number generation. Use
NULL for random number generation.}
}
\value{
Modified network.
}
\description{
This function takes a network and either adds or deletes a percentage of
edges randomly
}
