% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_eval}
\alias{run_eval}
\title{Run evaluation of results from decouple.}
\usage{
run_eval(
  bench_env,
  .design,
  res,
  .form = TRUE,
  .perform = TRUE,
  .downsample_pr = FALSE,
  .downsample_roc = FALSE,
  .downsample_times = 100,
  .silent = F
)
}
\arguments{
\item{.design}{design tibble.}

\item{res}{results from decouple.}

\item{.form}{bool whether to format or not}

\item{.perform}{bool whether to calculate ROC and performance summary}

\item{.downsample_pr}{whether to downsample precision recall curve TNs}

\item{.downsample_roc}{whether to downsample ROC true negatives}

\item{.downsample_times}{downsampling iterations}

\item{.silent}{Whether to run without verbose.}
}
\value{
An S4 object of \link{BenchResult-class}
}
\description{
Run evaluation of results from decouple.
}
