% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_methods}
\alias{run_methods}
\title{Run decouple on a design matrix.}
\usage{
run_methods(
  bench_env,
  .design,
  .minsize = 5,
  .silent = TRUE,
  .url_bool = FALSE
)
}
\arguments{
\item{.design}{design tibble.}

\item{.minsize}{Minsize to filter}

\item{.silent}{Whether to run without verbose.}

\item{.url_bool}{Wheter to download data.}
}
\value{
Tibble of activities.
}
\description{
Run decouple on a design matrix.
}
