% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{net_weight}
\alias{net_weight}
\title{This function takes a network and removes weight and/or directionality of the network}
\usage{
net_weight(network, weight_rm = "mor")
}
\arguments{
\item{network}{Network to modify.}

\item{weight_rm}{Columns from network specifying weights ('mor').}
}
\value{
Modified network.
}
\description{
This function takes a network and removes weight and/or directionality of the network
}
